requirements.txt
Django==2.2.5
django-model-utils==3.2.0
factory-boy==2.12.0
Faker==2.0.1
# pkg-resources==0.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
